integrating changes into clean lut branch
